Audit-Trail-System-Block-Chain-main/backend/requirements.txt
pymongo
python-dotenv
motor
fastapi
uvicorn
pydantic
openpyxl
pandas
python-multipart
xlrd
